refactor(provider): create normalize.lua, dissolve config_manager
Consolidate parameter normalization into `provider/normalize.lua`:
- `flatten_parameters()`, `resolve_max_tokens()`, `resolve_thinking()`
  + helpers moved from config_manager and base.lua

Dissolve `core/config/manager.lua` into `core.lua`:
- Validation + facade writes inlined in `initialize_provider()`
- `validate_parameters()` chain preserved (OpenAI override is live)
- Delete `core/config/` directory

Additional cleanup:
- Delete `metadata.default_parameters` from all providers + registry
- Make `try_import_from_buffer` static (no self param)
- Remove dead `frontmatter_overrides`/`frontmatter_override` from
  status collectors and display code
- Update stale `config_manager.prepare_config()` doc reference
