+++ tests/flemma/autopilot_spec.lua
@@ -865,7 +865,7 @@ describe("Autopilot integration", function()
     package.loaded["flemma.state"] = nil
     package.loaded["flemma.tools"] = nil
     package.loaded["flemma.core"] = nil
-    package.loaded["flemma.core.config.manager"] = nil
+    package.loaded["flemma.provider.normalize"] = nil
     package.loaded["flemma.provider.registry"] = nil
     package.loaded["flemma.models"] = nil
     package.loaded["flemma.autopilot"] = nil
@@ -946,7 +946,7 @@ describe("Autopilot race condition: sync + async tool dispatch", function()
     package.loaded["flemma.tools.approval"] = nil
     package.loaded["flemma.tools.executor"] = nil
     package.loaded["flemma.core"] = nil
-    package.loaded["flemma.core.config.manager"] = nil
+    package.loaded["flemma.provider.normalize"] = nil
     package.loaded["flemma.provider.registry"] = nil
     package.loaded["flemma.models"] = nil
     package.loaded["flemma.autopilot"] = nil
