+++ tests/flemma/frontmatter_eval_spec.lua
@@ -15,7 +15,7 @@ describe("Frontmatter evaluation caching", function()
     package.loaded["flemma.state"] = nil
     package.loaded["flemma.tools"] = nil
     package.loaded["flemma.core"] = nil
-    package.loaded["flemma.core.config.manager"] = nil
+    package.loaded["flemma.provider.normalize"] = nil
     package.loaded["flemma.provider.registry"] = nil
     package.loaded["flemma.models"] = nil
     package.loaded["flemma.autopilot"] = nil
