+++ tests/flemma/provider_openai_spec.lua
@@ -1241,8 +1241,9 @@ describe("OpenAI Provider", function()
       assert.is_true(openai.metadata.capabilities.outputs_thinking)
     end)
 
-    it("should include reasoning_summary in default parameters", function()
-      assert.equals("auto", openai.metadata.default_parameters.reasoning_summary)
+    it("should include reasoning_summary default in config_schema", function()
+      local defaults = openai.metadata.config_schema:materialize()
+      assert.equals("auto", defaults.reasoning_summary)
     end)
   end)
 end)
