+++ tests/flemma/resolve_thinking_spec.lua
@@ -1,9 +1,9 @@
-describe("base.resolve_thinking", function()
-  local base
+describe("normalize.resolve_thinking", function()
+  local normalize
 
   before_each(function()
-    package.loaded["flemma.provider.base"] = nil
-    base = require("flemma.provider.base")
+    package.loaded["flemma.provider.normalize"] = nil
+    normalize = require("flemma.provider.normalize")
   end)
 
   describe("budget-based provider (Anthropic-like)", function()
@@ -16,115 +16,115 @@ describe("base.resolve_thinking", function()
     }
 
     it("thinking_budget takes priority over unified thinking", function()
-      local result = base.resolve_thinking({ thinking_budget = 8192 }, caps)
+      local result = normalize.resolve_thinking({ thinking_budget = 8192 }, caps)
       assert.is_true(result.enabled)
       assert.are.equal(8192, result.budget)
       assert.are.equal("medium", result.level)
     end)
 
     it("thinking_budget takes priority even when thinking is also set", function()
-      local result = base.resolve_thinking({ thinking_budget = 8192, thinking = "high" }, caps)
+      local result = normalize.resolve_thinking({ thinking_budget = 8192, thinking = "high" }, caps)
       assert.is_true(result.enabled)
       assert.are.equal(8192, result.budget)
       assert.are.equal("medium", result.level)
     end)
 
     it("thinking_budget=0 disables", function()
-      local result = base.resolve_thinking({ thinking_budget = 0 }, caps)
+      local result = normalize.resolve_thinking({ thinking_budget = 0 }, caps)
       assert.is_false(result.enabled)
       assert.is_nil(result.level)
     end)
 
     it("thinking='max' maps to 32768", function()
-      local result = base.resolve_thinking({ thinking = "max" }, caps)
+      local result = normalize.resolve_thinking({ thinking = "max" }, caps)
       assert.is_true(result.enabled)
       assert.are.equal(32768, result.budget)
       assert.are.equal("max", result.level)
     end)
 
     it("thinking='high' maps to 16384", function()
-      local result = base.resolve_thinking({ thinking = "high" }, caps)
+      local result = normalize.resolve_thinking({ thinking = "high" }, caps)
       assert.is_true(result.enabled)
       assert.are.equal(16384, result.budget)
       assert.are.equal("high", result.level)
     end)
 
     it("thinking='medium' maps to 8192", function()
-      local result = base.resolve_thinking({ thinking = "medium" }, caps)
+      local result = normalize.resolve_thinking({ thinking = "medium" }, caps)
       assert.is_true(result.enabled)
       assert.are.equal(8192, result.budget)
       assert.are.equal("medium", result.level)
     end)
 
     it("thinking='low' maps to 2048", function()
-      local result = base.resolve_thinking({ thinking = "low" }, caps)
+      local result = normalize.resolve_thinking({ thinking = "low" }, caps)
       assert.is_true(result.enabled)
       assert.are.equal(2048, result.budget)
       assert.are.equal("low", result.level)
     end)
 
     it("thinking='minimal' clamps to min 1024", function()
-      local result = base.resolve_thinking({ thinking = "minimal" }, caps)
+      local result = normalize.resolve_thinking({ thinking = "minimal" }, caps)
       assert.is_true(result.enabled)
       assert.are.equal(1024, result.budget)
       assert.are.equal("low", result.level)
     end)
 
     it("thinking=500 clamps to min 1024", function()
-      local result = base.resolve_thinking({ thinking = 500 }, caps)
+      local result = normalize.resolve_thinking({ thinking = 500 }, caps)
       assert.is_true(result.enabled)
       assert.are.equal(1024, result.budget)
       assert.are.equal("low", result.level)
     end)
 
     it("thinking=5000 (numeric) uses exact value", function()
-      local result = base.resolve_thinking({ thinking = 5000 }, caps)
+      local result = normalize.resolve_thinking({ thinking = 5000 }, caps)
       assert.is_true(result.enabled)
       assert.are.equal(5000, result.budget)
       assert.are.equal("medium", result.level)
     end)
 
     it("thinking=false disables", function()
-      local result = base.resolve_thinking({ thinking = false }, caps)
+      local result = normalize.resolve_thinking({ thinking = false }, caps)
       assert.is_false(result.enabled)
       assert.is_nil(result.level)
     end)
 
     it("thinking=0 disables", function()
-      local result = base.resolve_thinking({ thinking = 0 }, caps)
+      local result = normalize.resolve_thinking({ thinking = 0 }, caps)
       assert.is_false(result.enabled)
       assert.is_nil(result.level)
     end)
 
     it("nil thinking means disabled", function()
-      local result = base.resolve_thinking({}, caps)
+      local result = normalize.resolve_thinking({}, caps)
       assert.is_false(result.enabled)
       assert.is_nil(result.level)
     end)
 
     it("floors float budget to integer", function()
-      local result = base.resolve_thinking({ thinking = 5000.7 }, caps)
+      local result = normalize.resolve_thinking({ thinking = 5000.7 }, caps)
       assert.is_true(result.enabled)
       assert.are.equal(5000, result.budget)
       assert.are.equal("medium", result.level)
     end)
 
     it("high thinking_budget maps to max level", function()
-      local result = base.resolve_thinking({ thinking_budget = 32768 }, caps)
+      local result = normalize.resolve_thinking({ thinking_budget = 32768 }, caps)
       assert.is_true(result.enabled)
       assert.are.equal(32768, result.budget)
       assert.are.equal("max", result.level)
     end)
 
     it("thinking_budget=16384 maps to high level", function()
-      local result = base.resolve_thinking({ thinking_budget = 16384 }, caps)
+      local result = normalize.resolve_thinking({ thinking_budget = 16384 }, caps)
       assert.is_true(result.enabled)
       assert.are.equal(16384, result.budget)
       assert.are.equal("high", result.level)
     end)
 
     it("low thinking_budget maps to low level", function()
-      local result = base.resolve_thinking({ thinking_budget = 1024 }, caps)
+      local result = normalize.resolve_thinking({ thinking_budget = 1024 }, caps)
       assert.is_true(result.enabled)
       assert.are.equal(1024, result.budget)
       assert.are.equal("low", result.level)
@@ -141,21 +141,21 @@ describe("base.resolve_thinking", function()
     }
 
     it("thinking='low' maps to 2048 (no clamping needed)", function()
-      local result = base.resolve_thinking({ thinking = "low" }, caps)
+      local result = normalize.resolve_thinking({ thinking = "low" }, caps)
       assert.is_true(result.enabled)
       assert.are.equal(2048, result.budget)
       assert.are.equal("low", result.level)
     end)
 
     it("thinking='minimal' maps to 128 (no clamping with min=1)", function()
-      local result = base.resolve_thinking({ thinking = "minimal" }, caps)
+      local result = normalize.resolve_thinking({ thinking = "minimal" }, caps)
       assert.is_true(result.enabled)
       assert.are.equal(128, result.budget)
       assert.are.equal("minimal", result.level)
     end)
 
     it("thinking=1 uses exact value (min is 1)", function()
-      local result = base.resolve_thinking({ thinking = 1 }, caps)
+      local result = normalize.resolve_thinking({ thinking = 1 }, caps)
       assert.is_true(result.enabled)
       assert.are.equal(1, result.budget)
       assert.are.equal("minimal", result.level)
@@ -171,82 +171,82 @@ describe("base.resolve_thinking", function()
     }
 
     it("reasoning takes priority over unified thinking", function()
-      local result = base.resolve_thinking({ reasoning = "high" }, caps)
+      local result = normalize.resolve_thinking({ reasoning = "high" }, caps)
       assert.is_true(result.enabled)
       assert.are.equal("high", result.effort)
       assert.are.equal("high", result.level)
     end)
 
     it("reasoning takes priority even when thinking is also set", function()
-      local result = base.resolve_thinking({ reasoning = "high", thinking = "low" }, caps)
+      local result = normalize.resolve_thinking({ reasoning = "high", thinking = "low" }, caps)
       assert.is_true(result.enabled)
       assert.are.equal("high", result.effort)
       assert.are.equal("high", result.level)
     end)
 
     it("thinking='medium' falls through when no reasoning", function()
-      local result = base.resolve_thinking({ thinking = "medium" }, caps)
+      local result = normalize.resolve_thinking({ thinking = "medium" }, caps)
       assert.is_true(result.enabled)
       assert.are.equal("medium", result.effort)
       assert.are.equal("medium", result.level)
     end)
 
     it("thinking=5000 maps to medium effort", function()
-      local result = base.resolve_thinking({ thinking = 5000 }, caps)
+      local result = normalize.resolve_thinking({ thinking = 5000 }, caps)
       assert.is_true(result.enabled)
       assert.are.equal("medium", result.effort)
       assert.are.equal("medium", result.level)
     end)
 
     it("thinking=1024 maps to low effort", function()
-      local result = base.resolve_thinking({ thinking = 1024 }, caps)
+      local result = normalize.resolve_thinking({ thinking = 1024 }, caps)
       assert.is_true(result.enabled)
       assert.are.equal("low", result.effort)
       assert.are.equal("low", result.level)
     end)
 
     it("thinking=32768 maps to max effort", function()
-      local result = base.resolve_thinking({ thinking = 32768 }, caps)
+      local result = normalize.resolve_thinking({ thinking = 32768 }, caps)
       assert.is_true(result.enabled)
       assert.are.equal("max", result.effort)
       assert.are.equal("max", result.level)
     end)
 
     it("thinking='max' passes through as max effort", function()
-      local result = base.resolve_thinking({ thinking = "max" }, caps)
+      local result = normalize.resolve_thinking({ thinking = "max" }, caps)
       assert.is_true(result.enabled)
       assert.are.equal("max", result.effort)
       assert.are.equal("max", result.level)
     end)
 
     it("thinking='minimal' passes through as minimal effort", function()
-      local result = base.resolve_thinking({ thinking = "minimal" }, caps)
+      local result = normalize.resolve_thinking({ thinking = "minimal" }, caps)
       assert.is_true(result.enabled)
       assert.are.equal("minimal", result.effort)
       assert.are.equal("minimal", result.level)
     end)
 
     it("thinking=100 maps to minimal effort", function()
-      local result = base.resolve_thinking({ thinking = 100 }, caps)
+      local result = normalize.resolve_thinking({ thinking = 100 }, caps)
       assert.is_true(result.enabled)
       assert.are.equal("minimal", result.effort)
       assert.are.equal("minimal", result.level)
     end)
 
     it("thinking=false disables", function()
-      local result = base.resolve_thinking({ thinking = false }, caps)
+      local result = normalize.resolve_thinking({ thinking = false }, caps)
       assert.is_false(result.enabled)
       assert.is_nil(result.level)
     end)
 
     it("nil thinking means disabled", function()
-      local result = base.resolve_thinking({}, caps)
+      local result = normalize.resolve_thinking({}, caps)
       assert.is_false(result.enabled)
       assert.is_nil(result.level)
     end)
 
     it("empty reasoning string falls through to thinking", function()
-      local result = base.resolve_thinking({ reasoning = "", thinking = "high" }, caps)
+      local result = normalize.resolve_thinking({ reasoning = "", thinking = "high" }, caps)
       assert.is_true(result.enabled)
       assert.are.equal("high", result.effort)
       assert.are.equal("high", result.level)
@@ -257,7 +257,7 @@ describe("base.resolve_thinking", function()
         local model_info = {
           thinking_effort_map = { minimal = "low", low = "low", medium = "medium", high = "high", max = "high" },
         }
-        local result = base.resolve_thinking({ thinking = "minimal" }, caps, model_info)
+        local result = normalize.resolve_thinking({ thinking = "minimal" }, caps, model_info)
         assert.is_true(result.enabled)
         assert.are.equal("low", result.effort)
         assert.are.equal("minimal", result.level)
@@ -267,7 +267,7 @@ describe("base.resolve_thinking", function()
         local model_info = {
           thinking_effort_map = { minimal = "low", low = "low", medium = "medium", high = "high", max = "xhigh" },
         }
-        local result = base.resolve_thinking({ thinking = "max" }, caps, model_info)
+        local result = normalize.resolve_thinking({ thinking = "max" }, caps, model_info)
         assert.is_true(result.enabled)
         assert.are.equal("xhigh", result.effort)
         assert.are.equal("max", result.level)
@@ -278,7 +278,7 @@ describe("base.resolve_thinking", function()
           thinking_effort_map = { minimal = "low", low = "low", medium = "medium", high = "high", max = "high" },
         }
         -- 100 maps to "minimal" via budget_to_effort, then "minimal" -> "low" via effort map
-        local result = base.resolve_thinking({ thinking = 100 }, caps, model_info)
+        local result = normalize.resolve_thinking({ thinking = 100 }, caps, model_info)
         assert.is_true(result.enabled)
         assert.are.equal("low", result.effort)
         assert.are.equal("minimal", result.level)
@@ -288,14 +288,14 @@ describe("base.resolve_thinking", function()
         local model_info = {
           thinking_effort_map = { minimal = "low", low = "low", medium = "medium", high = "high", max = "xhigh" },
         }
-        local result = base.resolve_thinking({ reasoning = "minimal" }, caps, model_info)
+        local result = normalize.resolve_thinking({ reasoning = "minimal" }, caps, model_info)
         assert.is_true(result.enabled)
         assert.are.equal("low", result.effort)
         assert.are.equal("minimal", result.level)
       end)
 
       it("falls back to raw effort when no effort map", function()
-        local result = base.resolve_thinking({ thinking = "minimal" }, caps, nil)
+        local result = normalize.resolve_thinking({ thinking = "minimal" }, caps, nil)
         assert.is_true(result.enabled)
         assert.are.equal("minimal", result.effort)
         assert.are.equal("minimal", result.level)
@@ -318,13 +318,13 @@ describe("base.resolve_thinking", function()
         min_thinking_budget = 512,
         max_thinking_budget = 24576,
       }
-      local result = base.resolve_thinking({ thinking = "minimal" }, caps, model_info)
+      local result = normalize.resolve_thinking({ thinking = "minimal" }, caps, model_info)
       assert.is_true(result.enabled)
       assert.are.equal(512, result.budget)
     end)
 
     it("falls back to hardcoded budgets when model_info is nil", function()
-      local result = base.resolve_thinking({ thinking = "minimal" }, caps, nil)
+      local result = normalize.resolve_thinking({ thinking = "minimal" }, caps, nil)
       assert.is_true(result.enabled)
       assert.are.equal(1024, result.budget) -- clamped to caps.min_thinking_budget
     end)
@@ -335,7 +335,7 @@ describe("base.resolve_thinking", function()
         min_thinking_budget = 512,
         max_thinking_budget = 24576,
       }
-      local result = base.resolve_thinking({ thinking = 50000 }, caps, model_info)
+      local result = normalize.resolve_thinking({ thinking = 50000 }, caps, model_info)
       assert.is_true(result.enabled)
       assert.are.equal(24576, result.budget)
     end)
@@ -346,7 +346,7 @@ describe("base.resolve_thinking", function()
         min_thinking_budget = 512,
         max_thinking_budget = 24576,
       }
-      local result = base.resolve_thinking({ thinking = 100 }, caps, model_info)
+      local result = normalize.resolve_thinking({ thinking = 100 }, caps, model_info)
       assert.is_true(result.enabled)
       assert.are.equal(512, result.budget)
     end)
@@ -357,7 +357,7 @@ describe("base.resolve_thinking", function()
         min_thinking_budget = 512,
         max_thinking_budget = 24576,
       }
-      local result = base.resolve_thinking({ thinking = "max" }, caps, model_info)
+      local result = normalize.resolve_thinking({ thinking = "max" }, caps, model_info)
       assert.is_true(result.enabled)
       assert.are.equal(24576, result.budget)
     end)
@@ -368,7 +368,7 @@ describe("base.resolve_thinking", function()
         min_thinking_budget = 512,
         max_thinking_budget = 24576,
       }
-      local result = base.resolve_thinking({ thinking_budget = 99999 }, caps, model_info)
+      local result = normalize.resolve_thinking({ thinking_budget = 99999 }, caps, model_info)
       assert.is_true(result.enabled)
       assert.are.equal(24576, result.budget)
     end)
@@ -383,13 +383,13 @@ describe("base.resolve_thinking", function()
     }
 
     it("always returns disabled", function()
-      local result = base.resolve_thinking({ thinking = "high" }, caps)
+      local result = normalize.resolve_thinking({ thinking = "high" }, caps)
       assert.is_false(result.enabled)
       assert.is_nil(result.level)
     end)
 
     it("returns disabled even with thinking_budget", function()
-      local result = base.resolve_thinking({ thinking_budget = 4096 }, caps)
+      local result = normalize.resolve_thinking({ thinking_budget = 4096 }, caps)
       assert.is_false(result.enabled)
       assert.is_nil(result.level)
     end)
