+++ tests/flemma/tool_concurrency_spec.lua
@@ -9,7 +9,7 @@ package.loaded["flemma.tools"] = nil
 package.loaded["flemma.tools.executor"] = nil
 package.loaded["flemma.tools.registry"] = nil
 package.loaded["flemma.core"] = nil
-package.loaded["flemma.core.config.manager"] = nil
+package.loaded["flemma.provider.normalize"] = nil
 package.loaded["flemma.bridge"] = nil
 package.loaded["flemma.provider.registry"] = nil
 package.loaded["flemma.models"] = nil
